fix (users API): update API error, success, and comments for users API

+++ app/api/users/route.js
@@ -1,8 +1,10 @@
-// route.js
-// Contains APIs for user account management
-// Get all users
-// Delete user
-//
+/* 
+    route.js
+    -----
+    Contains APIs for user accounts management
+    - Get all accounts
+    - Delete account
+*/
 
 // route: /api/users
 
@@ -13,57 +15,85 @@
 
 */
 import pool from '@/services/database';
+import { apiError, apiSuccess } from '@/services/response';
 
 /*
-    GET API endpoint
-    Queries all user data from users table
-    Returns JSON object array of all users
-    Error if request failed
+    GET
+    /api/users
+
+    Gets account information for all users
+
+    Parameters:
+    - none
+
+    Behavior:
+    - Gets all user data from the database
+    - Returns a JSON object containing rows of all users if successful
+    - Returns an error if GET fails
 */
 export async function GET() {
     console.log("get... ");
     try {
         const result = await pool.query('SELECT * FROM users');
-        // Create 
-        return Response.json(result.rows);
+
+        return apiSuccess(
+            "Successfully fetched all user data",
+            result.rows,
+            200
+        )
     } catch (error) {
-        console.error("err: ", error);
-        return Response.json("Failed to get: ", error);
+        return apiError(
+            "Failed to process GET request: " + error.message,
+            500
+        )
     }
 }
 
 /*
-    DELETE API endpoint
-    Parameter of HTTP request object with user ID as query 
-    Queries delete user with matching ID from db
-    Returns JSON object of deleted user if exists
-    Error if ID does not exist, or request failed
+    DELETE
+    /api/users?id={id}
 
-    Expected query parameter:
-        "id": value as integer
+    Deletes account with matching user ID
+
+    Function Parameter:
+    - req: HTTP request object containing method, URL, query params, body 
+
+    Query Parameter:
+    - id: ID of the user whose account will be deleted
+
+    Behavior:
+    - Deletes the user with matching ID from the database
+    - Returns the deleted user's account data if successful
+    - Returns an error if the user does not exist or if DELETE fails
 */
 export async function DELETE(req) {
     try {
         console.log("deleting...")
         const { searchParams } = new URL(req.url);
         const id = searchParams.get('id');
+        if (!id) {
+            return apiError(
+                "Missing parameter ID",
+                400
+            )
+        }
         const deleteUser = await pool.query('DELETE FROM users WHERE ID= $1 RETURNING *', [id]);
         console.log(id);
         if (deleteUser.rows.length === 0) {
-            return Response.json({
-                error: `User id ${id} does not exist`
-            })
+            return apiError(
+                `User ${id} does not exist`,
+                404
+            )
         }
-        const allUsers = await pool.query('SELECT * FROM users');
-
-        return Response.json({
-            delete: deleteUser.rows,
-            users: allUsers.rows
-        })
+        return apiSuccess(
+            "Successfully deleted user data",
+            deleteUser.rows,
+            200
+        )
     } catch (error) {
-        return Response.json({
-            error: `Failed to delete id: ${id}`,
-            msg: error.message
-        })
+        return apiError(
+            "Failed to process DELETE request: " + error.message,
+            500
+        )
     }
 }
